move generator to colorscope
